requirements.txt
# FPGA Design Optimization Agent Dependencies

# Core dependencies
mcp>=1.0.0

# OpenAI client (only needed for full agent mode using OpenRouter, not test mode)
openai>=1.0.0

# RapidWright MCP dependencies
rapidwright>=2025.2.1

# Vivado MCP dependencies (pexpect for process control)
pexpect>=4.8.0

# Context Manager dependencies
pyyaml>=6.0
tiktoken>=0.5.0
